adding transactions for save garbage

+++ src/main/java/com/co/latin/ecociudadano/repositories/ClientRepository.java
@@ -3,5 +3,5 @@ package com.co.latin.ecociudadano.repositories;
 import com.co.latin.ecociudadano.model.Client;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClientRepository extends CrudRepository<Client, Integer> {
+public interface ClientRepository extends CrudRepository<Client, Long> {
 }
